Refactoring: simplify package layout, add test structure
