+++ .github/workflows/python-tests.yaml
@@ -1,6 +1,6 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 
-name: Python Tests
+name: ci
 
 on:
   push:
@@ -12,6 +12,7 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+
     strategy:
       fail-fast: false
       matrix:
@@ -26,9 +27,13 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install flake8
+          python -m pip install flake8 pytest
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
           flake8 . --count --max-line-length=120 --show-source --statistics
+      - name: Test with pytest
+        run: |
+          python -m pytest
+        working-directory: ./python
