--- python/geoipsets/__init__.py
@@ -1,5 +1,0 @@
-# __init__.py
-
-# from .provider.maxmind import MaxMindProvider
-# from .provider.dbip import DbIpProvider
-# from .provider.provider import AddressFamily, Firewall
