--- python/geoipsets/provider/__init__.py
@@ -1,5 +1,0 @@
-# __init__.py
-
-# from .maxmind import MaxMindProvider
-# from .dbip import DbIpProvider
-# from .provider import Provider, AddressFamily, Firewall
